Revise `Submission.config`
our submission's config didn't actually save in db before. we just get
the default of it, so the update will got failed.
let `Submission.config` become a classmethod solve this problem, it will
query the db every time we want to get the submission's config.
also, this will add the ablility to hot reload submission config
without rebooting the service.