Correctly update high score (#126)
* Revise testcase about comment

* Set get pdf file timeout to 0

it seems that the flask default send_file_max_age is not 0, so users may
get old comment due to cache.
set the timeout to 0 in route `/<submission_id>/pdf/<item>`

* Update PDF file validation

* Remove submissions data

this file is generate by user submitting which should not commit to
version cocntrol

* Correctly update high score

in the previous version, the `high_scores` fields just compare the old
value and newest modified submission's score only. if we reduce the
submission's score which has the highest score, `high_score` will be the
old value.
the correct way to ensure the highest score is to query all submissions
in this problem submitted by the user.