Correctly update high score
in the previous version, the `high_scores` fields just compare the old
value and newest modified submission's score only. if we reduce the
submission's score which has the highest score, `high_score` will be the
old value.
the correct way to ensure the highest score is to query all submissions
in this problem submitted by the user.