Fix/handwritten (#136)
* Fix/email (#113)

* email to lowercase

* login email case insensitive

* fix error check when updating testcase (#114)

* Add test case for updating problem test case

* Use `replace` instead of `put` in `edit_problem_test_case`

if a test case exists, try to update it will casue a
`mongoengine.fields.GridFSError`, use `replace` to update a
existed file.

* Move original testcase to `default/` directory

* Revise `edit_problem_test_case`

just decide which function to call by whether the file exists.
no need to manually create a new file before `replace`.

* Add testcase for updating problem test case

* Add `BadTestCase` error for wrong file name in zip

if user upload invalid data, it will return 200 before, but user should
get 400 due to they provide a invalid test case.
add a exception for handling that,
and include excepted filenames in return data.

* Update CODEOWNERS

* Feat/handwritten (#116)

* add name check

* Handwritten submission grading now upgrade others

* Fix/handwritten permission (#115)

* add submission.permission()

* Students can not view other's handwritten submission score

* Fix wrongly querying submission's username

* Add test case to check handwritten score visibility

Co-authored-by: bogay <pojay11523@gmail.com>

* Update CODEOWNERS

* Feat/submission opt (#117)

* fix user's AC problems

* Add script for create users

* Save submission comment into db

for the sake of more convenient back up datas

* Add index for submission

* Revise submission filter

- query with non-exist user will return empty list now
- consider to performance, remove permission check in submission filter

* Let `Submission.get_code` return None on exception

the handled exception include
- KeyError: if the required file not exists in the submission zip file
- AttritubeError: if the code not uploaded
- UnicodeDecodeError: if the file content can not be decoded by UTF-8

* Set gunicorn default `loglevel` to debug

* Check submission comment existence

* Add testcase about uploading comment

Co-authored-by: skps2010 <skps2010030908@yahoo.com>

* Feat/pin ann (#118)

* add ann pytest

* make test_auth the first so it can pass

* Add `pinned` property for ann

Co-authored-by: bogay <pojay11523@gmail.com>

* Feat/handwritten (#119)

* add name check

* Handwritten submission grading now upgrade others

* *handwritten can only have 1 per problem now
*handwritten's score in student status can go down now
*I am sure teacher can see handwritten

* add test about handwritten submission
 replacing old one

* add handwritten score check

* delete ids from student status when remove
old handwritten submission

* fix submission id not in id list

* remove all previous submissions after submitting

* fix style

Co-authored-by: Uier <ap9940506@gmail.com>

* High score (#120)

* Add route to get highest score

* Add testcase for `/problem/<problem_id>/high-score`

* Return courses when viewing problem

* Can set `pinned` on creatin new ann (#121)

* Fix/problem testdata (#123)

* Return `caseCount` when get problem

* Change returned pdf filename for identifying

* Strip input/output in `Problem.detailed_info`

because the test data may be very large (10MB+), seperate these 2 data
to different for better performance.
also, the `limit` field is not available now.

* Revise test case due to route modification

Co-authored-by: bogay <pojay11523@gmail.com>

* Fix wrong behaviour in repeated comment (#125)

* Revise testcase about comment

* Set get pdf file timeout to 0

it seems that the flask default send_file_max_age is not 0, so users may
get old comment due to cache.
set the timeout to 0 in route `/<submission_id>/pdf/<item>`

* Update PDF file validation

* Correctly update high score (#126)

* Revise testcase about comment

* Set get pdf file timeout to 0

it seems that the flask default send_file_max_age is not 0, so users may
get old comment due to cache.
set the timeout to 0 in route `/<submission_id>/pdf/<item>`

* Update PDF file validation

* Remove submissions data

this file is generate by user submitting which should not commit to
version cocntrol

* Correctly update high score

in the previous version, the `high_scores` fields just compare the old
value and newest modified submission's score only. if we reduce the
submission's score which has the highest score, `high_score` will be the
old value.
the correct way to ensure the highest score is to query all submissions
in this problem submitted by the user.

* Fix rejudge & delete submission (#130)

* Revise testcase about comment

* Set get pdf file timeout to 0

it seems that the flask default send_file_max_age is not 0, so users may
get old comment due to cache.
set the timeout to 0 in route `/<submission_id>/pdf/<item>`

* Update PDF file validation

* Remove submissions data

this file is generate by user submitting which should not commit to
version cocntrol

* Correctly update high score

in the previous version, the `high_scores` fields just compare the old
value and newest modified submission's score only. if we reduce the
submission's score which has the highest score, `high_score` will be the
old value.
the correct way to ensure the highest score is to query all submissions
in this problem submitted by the user.

* Delete one submission also delete its files

- in the previous commit, the output, source code and comment will still
  exist in DB.
  override `Submission.delete` to delete these dependent files.
- if teacher haven't upload comment, user those who try to get comment
  will get a empty file. it doesn't make sense, backend should not
  response a file due to it didn't exist on server, return a
  404 instead.
- if `Submission.add`'s timestamp not passed, automatically use
  `datetime.now()`

* Fix `Submission.delete_ouput`

iterate `task.cases` instead of `task`, which will get the field names
in the document.

* Update rejudge test case

test the behaviour of rejudge many times instead of once

* submitting check pdf's file type now
* fix submitting ValidationError's response

* make checking PDF in zip faster

Co-authored-by: AlaRduTP <alardutp37259418@gmail.com>
Co-authored-by: Bogay <pojay11523@gmail.com>
Co-authored-by: Siao,Yu-Jie <a6194315535@gmail.com>
Co-authored-by: Uier <ap9940506@gmail.com>